dicom-ingestion-to-s3-healthimaging/requirements.txt
aws-cdk-lib>=2.0.0
constructs>=10.0.0
aws-cdk.aws-codestar-alpha>=2.0.0alpha1
aws-cdk.aws-lambda-python-alpha==2.26.0a0
colorama>=0.4.3
botocore>=1.20.0,<2.0.0
boto3>=1.17.0,<2.0.0